* commit '04fb8810881b916fab7bf7736fe00ab4b1fe7359':
  IFS-9424: Incoporated review comments
  IFS-9424: Fixed warnings

+++ ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/competitionsetup/transactional/CompetitionSetupTemplateServiceImplIntegrationTest.java
@@ -38,7 +38,7 @@ public class CompetitionSetupTemplateServiceImplIntegrationTest extends BaseAuth
 
         stream(CompetitionTypeEnum.values()).forEach(type -> {
             Competition competition = competitionRepository.save(newCompetition()
-                    .withId(null)
+                    .withId((Long) null)
                     .withFundingType(FundingType.GRANT)
                     .withFundingRules(FundingRules.SUBSIDY_CONTROL)
                     .build());
@@ -57,7 +57,7 @@ public class CompetitionSetupTemplateServiceImplIntegrationTest extends BaseAuth
 
         stream(FundingType.values()).forEach(type -> {
             Competition competition = competitionRepository.save(newCompetition()
-                    .withId(null)
+                    .withId((Long) null)
                     .withFundingType(type)
                     .withFundingRules(FundingRules.SUBSIDY_CONTROL)
                     .build());
@@ -77,7 +77,7 @@ public class CompetitionSetupTemplateServiceImplIntegrationTest extends BaseAuth
 
         stream(FundingRules.values()).forEach(fundingRules -> {
             Competition competition = competitionRepository.save(newCompetition()
-                    .withId(null)
+                    .withId((Long) null)
                     .withFundingType(FundingType.GRANT)
                     .withFundingRules(fundingRules)
                     .build());
